python/test-webhooks/requirements.txt
verifit >= 5.0.2

python-dotenv
flask
requests
